+++ packages/sources/anyblock/src/endpoint/index.ts
@@ -1,7 +1,5 @@
 import type { TInputParameters as GasPriceInputParameters } from './gasprice'
-import type { TInputParameters as VwapInputParameters } from './vwap'
 
-export type TInputParameters = GasPriceInputParameters | VwapInputParameters
+export type TInputParameters = GasPriceInputParameters
 
 export * as gasprice from './gasprice'
-export * as vwap from './vwap'