+++ packages/sources/cryptoapis-v2/src/config/index.ts
@@ -27,7 +27,7 @@ export function isCoinType(key: string): key is keyof typeof BLOCKCHAIN_NAME_BY_
   return !!BLOCKCHAIN_NAME_BY_TICKER[key.toLowerCase() as BlockchainTickers]
 }
 export const CHAIN_KEYS = ['mainnet', 'testnet'] as const
-export type ChainType = typeof CHAIN_KEYS[number]
+export type ChainType = (typeof CHAIN_KEYS)[number]
 export function isChainType(key: string): key is ChainType {
   return CHAIN_KEYS.includes(key as ChainType)
 }