+++ packages/sources/cryptocompare-test/src/endpoint/vwap.ts
@@ -1,7 +1,13 @@
 import { customSettings } from '../config'
-import { PriceEndpoint, PriceEndpointParams } from '@chainlink/external-adapter-framework/adapter'
+import {
+  PriceEndpoint,
+  PriceEndpointInputParameters,
+  PriceEndpointParams,
+} from '@chainlink/external-adapter-framework/adapter'
 import { SingleNumberResultResponse } from '@chainlink/external-adapter-framework/util'
 import { HttpTransport } from '@chainlink/external-adapter-framework/transports'
+import { InputParameters } from '@chainlink/external-adapter-framework/validation'
+import overrides from '../config/overrides.json'
 
 export const inputParams = {
   base: {
@@ -21,7 +27,7 @@ export const inputParams = {
     type: 'number',
     default: 24,
   },
-} as const
+} satisfies InputParameters & PriceEndpointInputParameters
 
 interface ResponseSchema {
   [quoteSymbol: string]: number
@@ -98,4 +104,5 @@ export const endpoint = new PriceEndpoint<BatchEndpointTypes>({
   aliases: ['crypto-vwap'],
   transport: httpTransport,
   inputParameters: inputParams,
+  overrides: overrides.cryptocompare,
 })