+++ packages/sources/elwood/src/index.ts
@@ -1,5 +1,5 @@
 import { expose, ServerInstance } from '@chainlink/external-adapter-framework'
-import { PriceAdapter } from '@chainlink/external-adapter-framework/adapter'
+import { CryptoPriceEndpoint, PriceAdapter } from '@chainlink/external-adapter-framework/adapter'
 import { cryptoEndpoint } from './endpoint'
 import { customSettings } from './config'
 
@@ -7,7 +7,7 @@ export const adapter = new PriceAdapter({
   name: 'ELWOOD',
   defaultEndpoint: 'crypto',
   customSettings,
-  endpoints: [cryptoEndpoint],
+  endpoints: [cryptoEndpoint as CryptoPriceEndpoint<any>],
 })
 
 export const server = (): Promise<ServerInstance | undefined> => expose(adapter)