+++ packages/sources/finage-test/src/endpoint/forex-router.ts
@@ -2,8 +2,8 @@ import { customSettings } from '../config'
 import { httpTransport } from './http/forex'
 import { PriceEndpoint, PriceEndpointParams } from '@chainlink/external-adapter-framework/adapter'
 import { SingleNumberResultResponse } from '@chainlink/external-adapter-framework/util'
-import { RoutingTransport } from '@chainlink/external-adapter-framework/transports/meta'
 import { wsTransport } from './ws/forex-ws'
+import overrides from '../config/overrides.json'
 
 export const inputParameters = {
   base: {
@@ -27,9 +27,11 @@ interface ResponseSchema {
   timestamp: number
 }
 
+export type ForexEndpointParams = PriceEndpointParams
+
 export type EndpointTypes = {
   Request: {
-    Params: PriceEndpointParams
+    Params: ForexEndpointParams
   }
   Response: SingleNumberResultResponse
   CustomSettings: typeof customSettings
@@ -39,16 +41,13 @@ export type EndpointTypes = {
   }
 }
 
-export const routingTransport = new RoutingTransport<EndpointTypes>(
-  {
-    WS: wsTransport,
-    HTTP: httpTransport,
-  },
-  (_, adapterConfig) => (adapterConfig.WS_ENABLED ? 'WS' : 'HTTP'),
-)
-
 export const endpoint = new PriceEndpoint<EndpointTypes>({
   name: 'forex',
-  transport: routingTransport,
+  transports: {
+    ws: wsTransport,
+    rest: httpTransport,
+  },
+  defaultTransport: 'rest',
   inputParameters: inputParameters,
+  overrides: overrides.finage,
 })