+++ packages/sources/gsr/src/index.ts
@@ -1,12 +1,12 @@
 import { expose, ServerInstance } from '@chainlink/external-adapter-framework'
-import { PriceAdapter } from '@chainlink/external-adapter-framework/adapter'
+import { CryptoPriceEndpoint, PriceAdapter } from '@chainlink/external-adapter-framework/adapter'
 import { customSettings } from './config'
 import { price } from './endpoint'
 
 export const adapter = new PriceAdapter({
   defaultEndpoint: 'price',
   name: 'GSR',
-  endpoints: [price],
+  endpoints: [price as CryptoPriceEndpoint<any>],
   customSettings,
 })