+++ packages/sources/gsr/test/integration/setup.ts
@@ -3,7 +3,7 @@ import { ServerInstance } from '@chainlink/external-adapter-framework'
 import { SuperTest, Test } from 'supertest'
 import { WebSocketClassProvider } from '@chainlink/external-adapter-framework/transports'
 import { Server, WebSocket } from 'mock-socket'
-import { PriceAdapter } from '@chainlink/external-adapter-framework/adapter'
+import { CryptoPriceEndpoint, PriceAdapter } from '@chainlink/external-adapter-framework/adapter'
 import { endpoint } from '../../src/endpoint/price'
 import { SettingsMap } from '@chainlink/external-adapter-framework/config'
 import { customSettings } from '../../src/config'
@@ -61,9 +61,9 @@ export const mockWebSocketServer = (URL: string) => {
 
 export const createAdapter = (): PriceAdapter<SettingsMap> => {
   return new PriceAdapter({
-    name: 'test',
+    name: 'TEST',
     defaultEndpoint: 'price',
-    endpoints: [endpoint],
+    endpoints: [endpoint as CryptoPriceEndpoint<any>],
     customSettings,
   })
 }