+++ packages/sources/ncfx/src/endpoint/crypto.ts
@@ -6,7 +6,7 @@ import {
 } from '@chainlink/external-adapter-framework/util'
 import { WebSocketTransport } from '@chainlink/external-adapter-framework/transports'
 import {
-  PriceEndpoint,
+  CryptoPriceEndpoint,
   priceEndpointInputParameters,
   PriceEndpointParams,
 } from '@chainlink/external-adapter-framework/adapter'
@@ -41,7 +41,7 @@ export const cryptoTransport = new WebSocketTransport<EndpointTypes>({
         // Set up listener
         connection.on('message', (data: WebSocketRawData) => {
           const parsed = JSON.parse(data.toString())
-          if (parsed.Message === 'Succesfully Authenticated') {
+          if (parsed.Message === 'Successfully Authenticated') {
             logger.debug('Got logged in response, connection is ready')
             resolve()
           } else {
@@ -94,7 +94,7 @@ export const cryptoTransport = new WebSocketTransport<EndpointTypes>({
   },
 })
 
-export const cryptoEndpoint = new PriceEndpoint<EndpointTypes>({
+export const cryptoEndpoint = new CryptoPriceEndpoint<EndpointTypes>({
   name: 'crypto',
   transport: cryptoTransport,
   inputParameters: priceEndpointInputParameters,