+++ packages/sources/polkadot-balance/src/endpoint/balance.ts
@@ -63,6 +63,7 @@ const chunkArray = (addresses: string[], size: number): string[][] =>
     : [addresses]
 
 export class BalanceTransport implements Transport<EndpointTypes> {
+  name!: string
   responseCache!: ResponseCache<{
     Request: EndpointTypes['Request']
     Response: EndpointTypes['Response']
@@ -70,6 +71,7 @@ export class BalanceTransport implements Transport<EndpointTypes> {
 
   async initialize(dependencies: TransportDependencies<EndpointTypes>): Promise<void> {
     this.responseCache = dependencies.responseCache
+    this.name = 'default_single_transport'
   }
 
   async foregroundExecute(
@@ -132,7 +134,7 @@ export class BalanceTransport implements Transport<EndpointTypes> {
         providerIndicatedTimeUnixMs: undefined,
       },
     }
-    await this.responseCache.write([{ params: req.requestContext.data, response }])
+    await this.responseCache.write(this.name, [{ params: req.requestContext.data, response }])
     return response
   }
 }