+++ packages/sources/tiingo-test/src/endpoint/common/forex-router.ts
@@ -1,9 +1,11 @@
-import { RoutingTransport } from '@chainlink/external-adapter-framework/transports/meta'
-import { AdapterEndpoint, PriceEndpointParams } from '@chainlink/external-adapter-framework/adapter'
+import { AdapterEndpoint } from '@chainlink/external-adapter-framework/adapter'
 import { httpTransport } from '../http/forex'
 import { customSettings } from '../../config'
 import { SingleNumberResultResponse } from '@chainlink/external-adapter-framework/util'
 import { wsTransport } from '../ws/forex'
+import { InputParameters } from '@chainlink/external-adapter-framework/validation'
+import { RouterPriceEndpointParams } from '../../crypto-utils'
+import overrides from '../../config/overrides.json'
 
 const inputParameters = {
   base: {
@@ -18,7 +20,7 @@ const inputParameters = {
     type: 'string',
     description: 'The quote to convert to',
   },
-} as const
+} satisfies InputParameters
 
 interface ProviderResponseBody {
   ticker: string
@@ -32,7 +34,7 @@ interface ProviderResponseBody {
 
 export type ForexEndpointTypes = {
   Request: {
-    Params: PriceEndpointParams
+    Params: RouterPriceEndpointParams
   }
   Response: SingleNumberResultResponse
   CustomSettings: typeof customSettings
@@ -42,17 +44,14 @@ export type ForexEndpointTypes = {
   }
 }
 
-export const routingTransport = new RoutingTransport<ForexEndpointTypes>(
-  {
-    WS: wsTransport,
-    HTTP: httpTransport,
-  },
-  (_, adapterConfig) => (adapterConfig?.WS_ENABLED ? 'WS' : 'HTTP'),
-)
-
 export const endpoint = new AdapterEndpoint<ForexEndpointTypes>({
   name: 'forex',
   aliases: ['fx', 'commodities'],
-  transport: routingTransport,
+  transports: {
+    ws: wsTransport,
+    rest: httpTransport,
+  },
+  defaultTransport: 'rest',
   inputParameters: inputParameters,
+  overrides: overrides.tiingo,
 })