+++ packages/composites/augur/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @chainlink/augur-adapter
 
+## 2.0.12
+
+### Patch Changes
+
+- 65014014d: Upgraded typescript version to 4.9.5
+- Updated dependencies [b29509be0]
+- Updated dependencies [65014014d]
+  - @chainlink/ea-bootstrap@2.26.0
+  - @chainlink/sportsdataio-adapter@1.3.20
+  - @chainlink/therundown-adapter@1.3.28
+
 ## 2.0.11
 
 ### Patch Changes