+++ packages/sources/alongside/CHANGELOG.md
@@ -1,13 +1,19 @@
 # @chainlink/alongside-adapter
 
+## 1.1.1
+
+### Patch Changes
+
+- 65014014d: Bumped v3 framework version. Upgraded typescript to version 4.9.5
+
 ## 1.1.0
 
 ### Minor Changes
 
-- b5cccb5c6: Updated to allow generic rpc provider
+- 8e6e562f8: Updated to allow generic rpc provider
 
 ## 1.0.0
 
 ### Major Changes
 
-- d40add020: Add alongside EA v3
+- f99d90c89: Add alongside EA v3