+++ packages/sources/coinmarketcap-test/CHANGELOG.md
@@ -1,10 +1,16 @@
 # @chainlink/coinmarketcap-test-adapter
 
+## 1.0.2
+
+### Patch Changes
+
+- 65014014d: Bumped v3 framework version. Upgraded typescript to version 4.9.5
+
 ## 1.0.1
 
 ### Patch Changes
 
-- 70d0197d7: Add RETH symbol -> ID override
+- e303dd766: Add RETH symbol -> ID override
 
 ## 1.0.0