+++ packages/sources/cryptocompare-test/CHANGELOG.md
@@ -1,10 +1,16 @@
 # @chainlink/cryptocompare-test-adapter
 
+## 1.1.1
+
+### Patch Changes
+
+- 65014014d: Bumped v3 framework version. Upgraded typescript to version 4.9.5
+
 ## 1.1.0
 
 ### Minor Changes
 
-- e1ba01ac5: Update ea-framework version and transports, add vwap, marketcap, volume endpoints
+- 8b3c35545: Update ea-framework version and transports, add vwap, marketcap, volume endpoints
 
 ## 1.0.0