+++ packages/sources/polkadot-balance/CHANGELOG.md
@@ -1,7 +1,13 @@
 # @chainlink/polkadot-balance-adapter
 
+## 1.0.1
+
+### Patch Changes
+
+- 65014014d: Bumped v3 framework version. Upgraded typescript to version 4.9.5
+
 ## 1.0.0
 
 ### Major Changes
 
-- 603bb76e5: Initial version of EA
+- dee6d49a3: Initial version of EA