+++ packages/sources/tiingo-test/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @chainlink/tiingo-test-adapter
 
+## 1.0.1
+
+### Patch Changes
+
+- 65014014d: Bumped v3 framework version. Removed WS_ENABLED env var. Upgraded typescript to version 4.9.5
+
 ## 1.0.0
 
 ### Major Changes