+++ packages/sources/tradermade-test/CHANGELOG.md
@@ -1,7 +1,13 @@
 # @chainlink/tradermade-test-adapter
 
+## 1.0.1
+
+### Patch Changes
+
+- 65014014d: Bumped v3 framework version. Upgraded typescript to version 4.9.5
+
 ## 1.0.0
 
 ### Major Changes
 
-- 32e98dc5e: Initial version of the adapter
+- 49680b472: Initial version of the adapter