+++ packages/sources/twosigma/CHANGELOG.md
@@ -1,7 +1,14 @@
 # @chainlink/twosigma-adapter
 
+## 1.0.1
+
+### Patch Changes
+
+- 65014014d: Upgraded typescript version to 4.9.5
+- 65014014d: Bumped v3 framework version. Upgraded typescript to version 4.9.5
+
 ## 1.0.0
 
 ### Major Changes
 
-- bf4ea9f38: Add Twosigma external adapter
+- d1886603c: Add Twosigma external adapter