+++ packages/sources/wbtc-address-set/CHANGELOG.md
@@ -1,10 +1,21 @@
 # @chainlink/wbtc-address-set-adapter
 
+## 1.4.39
+
+### Patch Changes
+
+- 65014014d: Upgraded typescript version to 4.9.5
+- Updated dependencies [b29509be0]
+- Updated dependencies [65014014d]
+  - @chainlink/ea-bootstrap@2.26.0
+  - @chainlink/proof-of-reserves-adapter@1.16.1
+  - @chainlink/ea-test-helpers@1.4.2
+
 ## 1.4.38
 
 ### Patch Changes
 
-- Updated dependencies [603bb76e5]
+- Updated dependencies [dee6d49a3]
   - @chainlink/proof-of-reserves-adapter@1.16.0
 
 ## 1.4.37